pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fars-mvp"
version = "0.1.0"
description = "FARS MVP: unattended ideation -> experiment -> writing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.0",
  "jinja2>=3.1",
  "matplotlib>=3.7",
  "pyyaml>=6.0"
]

[tool.setuptools.packages.find]
where = ["src"]
